renaming and using format in configs
