+++ src/interfaces/queue.js
